Use babel-preset-env to determine if target browsers require polyfill
If the application has specified targets and those targets fully support
regenerator, don’t include the regenerator runtime.

This is not finished nor it will be until some work is done in ember-cli
and also the name of the transformations in babel5/6 might have changed,
so this is will probably only work in babel6.